chore: gate sqlite
