+++ clarity/Cargo.toml
@@ -27,7 +27,7 @@ regex = "1"
 lazy_static = "1.4.0"
 integer-sqrt = "0.1.3"
 slog = { version = "2.5.2", features = [ "max_level_trace" ] }
-stacks_common = { package = "stacks-common", path = "../stacks-common/." }
+stacks_common = { package = "stacks-common", path = "../stacks-common/.", optional = true, default-features = false }
 rstest = "0.11.0"
 rstest_reuse = "0.1.3"
 
@@ -36,8 +36,9 @@ version = "1.0"
 features = ["arbitrary_precision", "unbounded_depth"]
 
 [dependencies.rusqlite]
-version = "=0.24.2"
+version = "=0.28.0"
 features = ["blob", "serde_json", "i128_blob", "bundled", "trace"]
+optional = true
 
 [dependencies.time]
 version = "0.2.23"
