+++ src/chainstate/stacks/boot/contract_tests.rs
@@ -22,7 +22,6 @@ use crate::core::{
 };
 use crate::util_lib::db::{DBConn, FromRow};
 use clarity::vm::analysis::arithmetic_checker::ArithmeticOnlyChecker;
-use clarity::vm::analysis::mem_type_check;
 use clarity::vm::ast::ASTRules;
 use clarity::vm::contexts::OwnedEnvironment;
 use clarity::vm::contracts::Contract;
@@ -35,6 +34,7 @@ use clarity::vm::errors::{
 use clarity::vm::eval;
 use clarity::vm::representations::SymbolicExpression;
 use clarity::vm::test_util::{execute, symbols_from_values, TEST_BURN_STATE_DB, TEST_HEADER_DB};
+use clarity::vm::tooling::mem_type_check;
 use clarity::vm::types::Value::Response;
 use clarity::vm::types::{
     OptionalData, PrincipalData, QualifiedContractIdentifier, ResponseData, StandardPrincipalData,
