+++ stacks-common/Cargo.toml
@@ -41,8 +41,9 @@ version = "0.24.3"
 features = ["serde", "recovery"]
 
 [dependencies.rusqlite]
-version = "=0.24.2"
+version = "=0.28.0"
 features = ["blob", "serde_json", "i128_blob", "bundled", "trace"]
+optional = true
 
 [dependencies.ed25519-dalek]
 version = "=1.0.0-pre.3"
